refactored it kinda more back to how it was, but reduced the complexity significantly in class depth
