+++ grug/__main__.py
@@ -4,7 +4,7 @@ import anyio
 from loguru import logger
 
 from grug.db import init_db
-from grug.discord_client import start_discord_bot
+from grug.discord_client import DiscordClient
 from grug.scheduler import start_scheduler
 from grug.settings import settings
 
@@ -24,7 +24,7 @@ async def main():
     init_db()
 
     async with anyio.create_task_group() as tg:
-        tg.start_soon(start_discord_bot)
+        tg.start_soon(DiscordClient().start, settings.discord_token.get_secret_value())
         tg.start_soon(start_scheduler)
 
     logger.info("Grug has shut down...")
