--- grug/discord_client/__init__.py
@@ -1,6 +1,0 @@
-from grug.discord_client.client import get_bot_invite_url, start_discord_bot
-
-__all__ = [
-    "start_discord_bot",
-    "get_bot_invite_url",
-]
