+++ grug/discord_speech_recognition.py
@@ -1,21 +1,21 @@
 """
-Voice Agent used for TTS and STT in Discord voice channels.
+Discord Speech Recognition Sink
 
-Based on the [SpeechRecognitionSink](https://github.com/imayhaveborkedit/discord-ext-voice-recv/blob/main/discord/ext/voice_recv/extras/speechrecognition.py)
+based on the [SpeechRecognitionSink](https://github.com/imayhaveborkedit/discord-ext-voice-recv/blob/main/discord/ext/voice_recv/extras/speechrecognition.py)
 from discord-ext-voice-recv.
 """
 
 import array
 import asyncio
 import audioop
+import concurrent.futures
 import time
 from collections import defaultdict
-from concurrent.futures import Future as CFuture
 from datetime import UTC, datetime
-from typing import Awaitable, Final, Optional, Protocol, TypedDict, TypeVar
+from typing import Any, Awaitable, Final, Optional, TypedDict, TypeVar
 
-import speech_recognition as sr  # type: ignore
-from discord import User, VoiceChannel
+import discord
+import speech_recognition as sr
 from discord.ext.voice_recv import AudioSink, SilencePacket, VoiceData
 from loguru import logger
 from speech_recognition.recognizers.whisper_api import openai as sr_openai
@@ -24,101 +24,10 @@ from tembo_pgmq_python import PGMQueue
 from grug.settings import settings
 
 
-class SRStopper(Protocol):
-    def __call__(self, wait: bool = True, /) -> None: ...
-
-
 class _StreamData(TypedDict):
-    stopper: Optional[SRStopper]
+    stopper: Optional[Any]
     recognizer: sr.Recognizer
     buffer: array.array[int]
-
-
-class SpeechRecognitionSink(AudioSink):  # type: ignore
-    _stream_data: defaultdict[int, _StreamData] = defaultdict(
-        lambda: _StreamData(stopper=None, recognizer=sr.Recognizer(), buffer=array.array("B"))
-    )
-
-    def __init__(self, discord_channel: VoiceChannel):
-        super().__init__(None)
-        self.discord_channel: VoiceChannel = discord_channel
-
-        self.queue = PGMQueue(
-            host=settings.postgres_host,
-            port=settings.postgres_port,
-            username=settings.postgres_user,
-            password=settings.postgres_password.get_secret_value(),
-            database=settings.postgres_db,
-        )
-
-        # Create a queue for the voice channel if it doesn't exist
-        if str(discord_channel.id) not in self.queue.list_queues():
-            self.queue.create_queue(str(discord_channel.id))
-
-    def _await(self, coro: Awaitable[TypeVar]) -> CFuture[TypeVar]:
-        assert self.client is not None
-        return asyncio.run_coroutine_threadsafe(coro, self.client.loop)
-
-    def wants_opus(self) -> bool:
-        return False
-
-    def write(self, user: Optional[User], data: VoiceData) -> None:
-        # Ignore silence packets and packets from users we don't have data for
-        if isinstance(data.packet, SilencePacket) or user is None:
-            return
-
-        sdata = self._stream_data[user.id]
-        sdata["buffer"].extend(data.pcm)
-
-        if not sdata["stopper"]:
-            sdata["stopper"] = sdata["recognizer"].listen_in_background(
-                source=_DiscordSRAudioSource(sdata["buffer"]),
-                callback=self.background_listener(user),
-                phrase_time_limit=10,
-            )
-
-    def background_listener(self, user: User):
-        def callback(_recognizer: sr.Recognizer, _audio: sr.AudioData):
-            # Don't process empty audio data or audio data that is too small
-            if _audio.frame_data == b"" or len(bytes(_audio.frame_data)) < 10000:
-                return None
-
-            # Get the text from the audio data
-            text_output = None
-            try:
-                text_output = sr_openai.recognize(_recognizer, _audio)
-            except sr.UnknownValueError:
-                logger.debug("Bad speech chunk")
-
-            # WEIRDEST BUG EVER: for some reason whisper keeps getting the word "you" from the recognizer, so
-            #                    we'll just ignore any text segments that are just "you"
-            if text_output and text_output.lower() != "you":
-                self.queue.send(
-                    str(self.discord_channel.id),
-                    {
-                        "user_id": user.id,
-                        "message_timestamp": datetime.now(tz=UTC).isoformat(),
-                        "message": text_output,
-                    },
-                )
-
-        return callback
-
-    def cleanup(self) -> None:
-        for user_id in tuple(self._stream_data.keys()):
-            self._drop(user_id)
-
-    def _drop(self, user_id: int) -> None:
-        if user_id in self._stream_data:
-            data = self._stream_data.pop(user_id)
-            stopper = data.get("stopper")
-            if stopper:
-                stopper()
-
-            buffer = data.get("buffer")
-            if buffer:
-                # arrays don't have a clear function
-                del buffer[:]
 
 
 class _DiscordSRAudioSource(sr.AudioSource):
@@ -167,3 +76,90 @@ class _DiscordSRAudioSource(sr.AudioSource):
 
     def close(self) -> None:
         self.buffer.clear()
+
+
+class SpeechRecognitionSink(AudioSink):
+    _stream_data: defaultdict[int, _StreamData] = defaultdict(
+        lambda: _StreamData(stopper=None, recognizer=sr.Recognizer(), buffer=array.array("B"))
+    )
+
+    def __init__(self, discord_channel: discord.VoiceChannel):
+        super().__init__(None)
+        self.discord_channel: discord.VoiceChannel = discord_channel
+
+        self.queue = PGMQueue(
+            host=settings.postgres_host,
+            port=settings.postgres_port,
+            username=settings.postgres_user,
+            password=settings.postgres_password.get_secret_value(),
+            database=settings.postgres_db,
+        )
+
+        # Create a queue for the voice channel if it doesn't exist
+        if str(discord_channel.id) not in self.queue.list_queues():
+            self.queue.create_queue(str(discord_channel.id))
+
+    def _await(self, coro: Awaitable[TypeVar]) -> concurrent.futures.Future[TypeVar]:
+        assert self.client is not None
+        return asyncio.run_coroutine_threadsafe(coro, self.client.loop)
+
+    def wants_opus(self) -> bool:
+        return False
+
+    def write(self, user: Optional[discord.User], data: VoiceData) -> None:
+        # Ignore silence packets and packets from users we don't have data for
+        if isinstance(data.packet, SilencePacket) or user is None:
+            return
+
+        sdata = self._stream_data[user.id]
+        sdata["buffer"].extend(data.pcm)
+
+        if not sdata["stopper"]:
+            sdata["stopper"] = sdata["recognizer"].listen_in_background(
+                source=_DiscordSRAudioSource(sdata["buffer"]),
+                callback=self.background_listener(user),
+                phrase_time_limit=10,
+            )
+
+    def background_listener(self, user: discord.User):
+        def callback(_recognizer: sr.Recognizer, _audio: sr.AudioData):
+            # Don't process empty audio data or audio data that is too small
+            if _audio.frame_data == b"" or len(bytes(_audio.frame_data)) < 10000:
+                return None
+
+            # Get the text from the audio data
+            text_output = None
+            try:
+                text_output = sr_openai.recognize(_recognizer, _audio)
+            except sr.UnknownValueError:
+                logger.debug("Bad speech chunk")
+
+            # WEIRDEST BUG EVER: for some reason whisper keeps getting the word "you" from the recognizer, so
+            #                    we'll just ignore any text segments that are just "you"
+            if text_output and text_output.lower() != "you":
+                self.queue.send(
+                    str(self.discord_channel.id),
+                    {
+                        "user_id": user.id,
+                        "message_timestamp": datetime.now(tz=UTC).isoformat(),
+                        "message": text_output,
+                    },
+                )
+
+        return callback
+
+    def cleanup(self) -> None:
+        for user_id in tuple(self._stream_data.keys()):
+            self._drop(user_id)
+
+    def _drop(self, user_id: int) -> None:
+        if user_id in self._stream_data:
+            data = self._stream_data.pop(user_id)
+            stopper = data.get("stopper")
+            if stopper:
+                stopper()
+
+            buffer = data.get("buffer")
+            if buffer:
+                # arrays don't have a clear function
+                del buffer[:]
